Move countries to util/countries-data.
